% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkyWiseNetCDF.R
\name{ExtractGrid}
\alias{ExtractGrid}
\title{ExtractGrid}
\usage{
ExtractGrid(fileName, var)
}
\arguments{
\item{fileName}{the netcdf filename you want to extract a variable from}

\item{var}{the name of the variable you want to extract from the netcdf file}
}
\value{
a named list containing the variable and the file attributes
}
\description{
Extracts a parameter from the netcdf as well as the netcdf attributes
}
\examples{
\dontrun{ grid <- ExtractGrid("file.nc", "temperature") }
}
