% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkyWiseNetCDF.R
\name{GetValueAtPoint}
\alias{GetValueAtPoint}
\title{GetValueAtPoint}
\usage{
GetValueAtPoint(search_lat, search_lon, grid)
}
\arguments{
\item{search_lat}{the latitude to find}

\item{search_lon}{the longitude to find}

\item{grid}{the grid containing the variable retrieved from the ExtractGrid call}
}
\value{
the value contained in the grid at the closest lat/lon matching point
}
\description{
Returns the value of a grid at a given lat/lon
}
\examples{
\dontrun{ val <- GetValueAtPoint(35, -97, grid) }
}
